ZK-112: Move row selection state into JournalSliceContext
