+++ web/src/components/journal/JournalEditor.tsx
@@ -20,7 +20,6 @@ export default function JournalEditor() {
 		entry: null,
 		anchorEl: null,
 	})
-	const [selectedRows, setSelectedRows] = useState<Record<string, boolean>>({})
 
 	const { snackbar } = useContext(NotificationsContext)
 	const journalContext = useContext(JournalContext)
@@ -99,61 +98,7 @@ export default function JournalEditor() {
 		}
 	}
 
-	const toggleSelectedRow = (row: string) => {
-		setSelectedRows((prev) => {
-			return {
-				...prev,
-				[row]: prev[row] ? false : true
-			}
-		})
-	}
-
-	const handleSelectAll = (action: SelectAllAction) => {
-		console.log(`handleSelectAll(${action})`)
-		setSelectedRows((prev) => {
-			let selected: Set<string>
-			const allRowIds = new Set<string>(Object.keys(journalSliceContext.getJournalEntriesQuery.data ?? {}))
-			const emptySet = new Set<string>([])
-			const hasSelectedAll = Object.values(prev).length > 0 && Object.values(prev).every(Boolean)
-
-			switch (action) {
-				case SelectAllAction.ALL:
-					selected = allRowIds
-					break
-
-				case SelectAllAction.NONE:
-					selected = emptySet
-					break
-
-				case SelectAllAction.CREDIT:
-					selected = new Set<string>(Array.from(allRowIds).filter((id: string) => {
-						const entry = journalSliceContext.getJournalEntriesQuery.data[id]
-						return entry ? calculateNetAmount(entry) > 0 : false
-					}))
-					break
-
-				case SelectAllAction.DEBIT:
-					selected = new Set<string>(Array.from(allRowIds).filter((id: string) => {
-						const entry = journalSliceContext.getJournalEntriesQuery.data[id]
-						return entry ? calculateNetAmount(entry) < 0 : false
-					}))
-					break
-
-				case SelectAllAction.TOGGLE:
-				default:
-					selected = hasSelectedAll ? emptySet : allRowIds
-			}
-
-			return Object.fromEntries(Array.from(new Set([...Object.keys(prev), ...selected]))
-				.map((key) => {
-					return [key, selected.has(key)]
-				}))
-		})
-	}
-
-	useEffect(() => {
-		setSelectedRows({})
-	}, [journalSliceContext.getJournalEntriesQuery.data])
+	
 
 	// show all docs
 	useEffect(() => {
@@ -181,19 +126,13 @@ export default function JournalEditor() {
 					overflow: 'hidden',
 					flex: 1,
 				})}>
-				<JournalHeader
-					numRows={Object.values(journalSliceContext.getJournalEntriesQuery.data ?? {}).length}
-					numSelectedRows={Object.values(selectedRows).filter(Boolean).length}
-					onSelectAll={handleSelectAll}
-				/>
+				<JournalHeader />
 				<Divider />
 				<Box sx={{
 					flex: 1,
 					overflowY: 'auto',
 				}}>
 					<JournalEntryList
-						selectedRows={selectedRows}
-						toggleSelectedRow={toggleSelectedRow}
 						journalRecordGroups={journalGroups}
 						onClickListItem={handleClickListItem}
 						onDoubleClickListItem={handleDoubleClickListItem}
