+++ web/src/contexts/JournalSliceContext.tsx
@@ -1,3 +1,4 @@
+import { SelectAllAction } from '@/components/journal/JournalHeader'
 import { DateView, JournalEntry } from '@/types/schema'
 import { DefinedUseQueryResult } from '@tanstack/react-query'
 import { createContext } from 'react'
@@ -21,6 +22,12 @@ interface JournalSliceContext extends JournalEditorState {
 		Error
 	>
 	refetchAllDependantQueries: () => void
+
+	// Selecting
+	numRows: number
+	selectedRows: Record<string, boolean>
+	onSelectAll: (action: SelectAllAction) => void
+	toggleSelectedRow: (row: string) => void
 }
 
 export const JournalSliceContext = createContext<JournalSliceContext>({} as JournalSliceContext)
